Add Sorting2 project with Merge Sort implementation
This commit introduces a new project "Sorting2" to the solution, featuring an implementation of the Merge Sort algorithm within the `MergeSort` class. The `Program.cs` file has been updated to demonstrate the sorting functionality. Additionally, several LeetCode problem classes have been added or modified, including algorithms for checking if an array is sorted, rotating an array, and determining the maximum frequency of an element. The project file for "Sorting2" is configured for .NET 9.0.